change BOW generator, sha1 → blake2, separate function for file names
