+++ trafilatura/cli_utils.py
@@ -24,7 +24,7 @@ from trafilatura import spider
 
 from .core import extract
 from .downloads import add_to_compressed_dict, buffered_downloads, load_download_buffer
-from .filters import content_fingerprint
+from .filters import generate_hash_filename
 from .utils import uniquify_list
 from .settings import (use_config, FILENAME_LEN,
                        FILE_PROCESSING_CORES, MAX_FILES_PER_DIRECTORY)
@@ -141,7 +141,7 @@ def determine_output_path(args, orig_filename, content, counter=None, new_filena
         destination_directory = determine_counter_dir(args.output_dir, counter)
         # use cryptographic hash on file contents to define name
         if new_filename is None:
-            new_filename = content_fingerprint(content)[:27].replace('/', '-')
+            new_filename = generate_hash_filename(content)
         output_path = path.join(destination_directory, new_filename + extension)
     return output_path, destination_directory
